==> code/1_siteComparisonEnvironmental.Rmd <==
---
title: "1_SiteComparisonEnvironmental"
author: "Sara Devi Swaminathan"
date: "`r Sys.Date()`"
output: html_document
toc: true
toc_float: true
number_sections: true
---

# Load packages and set working directory

```{r setup, include=FALSE}
library(tidyverse)  # Comprehensive data manipulation and visualization
library(readxl)     # Excel file reading
library(zoo)        # Time series manipulation
library(parallel)   # Parallel computing
library(brms)       # Bayesian regression models
library(reshape2)   # Data manipulation
library(bayestestR) # Tools for Bayesian model checking and comparisons

# Set the root directory for knitr to the project root in an RStudio project.
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

# Read in and format YSI data
- Format columns and remove partial (deployment/retrieval) days

```{r}
#-Load data frame, format columns, and remove partial days-#
YSIfiltered <- read_xlsx("data/sonde_data_withSeptNurseryMinidot.xlsx") %>%                               # read in sonde data file
  mutate(date = as.Date(date, format = '%m/%d/%y')) %>%                                                   # reformat date
  mutate(time = as.character(gsub(".* ","",time))) %>%                                                    # remove date from time
  mutate(site = as.factor(site)) %>%                                                                      # make site a factor variable
  mutate(DT = as.POSIXct(paste(date, time), format = "%Y-%m-%d %H:%M:%S")) %>%                            # make datetime column
  mutate(fulldays = factor(ifelse(date >= "2019-06-12" & date <= "2019-07-07", "JJ",                      # Make column to mark different deployments; mark june-july deployment
                       ifelse(date >= "2019-07-10" & date <= "2019-08-04", "JA",                          # mark july-august deployment
                          ifelse(date >= "2019-08-07" & date <= "2019-09-08", "AS",                       # mark august-september deployment
                                  ifelse(date >= "2019-09-10" & date <= "2019-10-07", "SO", NA)))))) %>%  # mark september-october deployment, mark as NA for partial days
  mutate(fulldays = factor(fulldays, levels = c("JJ", "JA", "AS", "SO"))) %>%                             # make "fulldays" a factor variable
  filter(!is.na(fulldays))                                                                                # remove partial days
```

## Make moving averages and summary dataframes
- Calculate moving averages of temperature, dissolved oxygen, total dissolved solids, pH, and salinity
- Make a summary dataframe that summarizes across site and date

```{r}
#-Make a dataframe with 24-hour moving averages of each water quality paramter-#
YSI_ma <- YSIfiltered %>%                                          # take filtered YSI dataframe
  dplyr::group_by(site) %>%                                        # group by site 
  dplyr::arrange(DT, .by_group = TRUE) %>%                         # arrange by day and time within each site
  mutate(datetime_index = as.numeric(DT))%>%                       # make a numeric datetime column for models   
  mutate(maTemp_24 = rollmean(Temp_degC, k = 144, fill = NA),      # calculate 24-hr moving average of temperature (k = 24 hrs * 60 min = 1440 / measurements every 10 minutes)
    maDO_24 = rollmean(DOmgL, k = 144, fill = NA),                 # calculate 24-hr moving average of DO 
    maTDS_24 = rollmean(TDS_mgL, k = 144, fill = NA),              # calculate 24-hr moving average of TDS 
    mapH_24 = rollmean(pH, k = 144, fill = NA),                    # calculate 24-hr moving average of pH 
    maSal_24 = rollmean(Sal_PSU, k = 144, fill = NA))              # calculate 24-hr moving average of salinity 

#-Make a separate moving averages dataframe for munson (inshore) data-#
ma_munson <- YSI_ma %>%    # take moving averages dataframe
  filter(site == "munson") # filter to only munson (inshore) site

#-Make a separate moving averages dataframe for nursery (offshore) data-#
ma_nursery <- YSI_ma %>%    # take moving averages dataframe
  filter(site == "nursery") # filter to only nursery (offshore) site

#-Summarize moving average df with min and max values-#
ysiSumm <- YSI_ma %>%                   # take moving averages dataframe
  group_by(site, date)%>%               # group by site and date
  mutate(DOmax = max(DOmgL),            # calculate mean, max, and min for each response variable
            DOmin = min(DOmgL), 
            DOmean = mean(DOmgL), 
            tempMax = max(Temp_degC), 
            tempMin = min(Temp_degC), 
            tempMean = mean(Temp_degC), 
            pHmax = max(pH),
            pHmin = min(pH),
            pHmean = mean(pH), 
            Sal_PSUmax = max(Sal_PSU),
            Sal_PSUmin = min(Sal_PSU),
            Sal_PSUmean = mean(Sal_PSU),
            TDSMax = max(TDS_mgL), 
            TDSMin = min(TDS_mgL), 
            TDSMean = mean(TDS_mgL))

#-Subset to temperature data-#
YSI_ma_temp<- YSI_ma %>%                                            # read moving averages dataframe
  dplyr::select(Temp_degC, site, DT, datetime_index, maTemp_24) %>% # select only relevant columns
  unique()%>%                                                       # remove duplicate rows
  group_by(datetime_index, site)%>%                                 # group by datetime and site
  summarise(Temp_degC = mean(Temp_degC))%>%                         # ensure only one Temp reading for each datetime/site column
  mutate(logTemp = log(Temp_degC))                                  # calculate log(Temp) for models

#-Subset to DO data-#
YSI_ma_do<- YSI_ma %>%                                       # read moving averages dataframe
  dplyr::select(DOmgL, site, DT, datetime_index, maDO_24)%>% # select only relevant columns
  unique()%>%                                                # remove duplicate rows
  group_by(datetime_index, site)%>%                          # group by datetime and site
  summarise(DOmgL = mean(DOmgL))                             # ensure only one DO reading for each datetime/site column
```

# Summary statistics

## Inshore site summary statistics

```{r}
#-Calculate mean temperature +/- SE-#
ma_munson%>%                        # Munson/inshore site data
  filter(!is.na(maTemp_24)) %>%     # remove NAs
  summarise(N = n_distinct(date),   # get number of dates
            mean = mean(maTemp_24), # get mean temperature
            SD = sd(maTemp_24),     # calculate standard deviation
            SE = SD/sqrt(N))        # calculate standard error

#-Calculate # days where daily temperature mean was >31 degC inshore-#
ma_munson%>%                                 # Munson/inshore site data
  ungroup()%>%                               # remove any groupings
  dplyr::group_by(date)%>%                   # group by date
  summarise(dailyMean = mean(maTemp_24))%>%  # calculate mean Temp for each date
  filter(dailyMean > 31)                     # filter to only days when mean > 31degC

#-Calculate mean DO +/- SE-#
ma_munson%>%                     # Munson/inshore site data
  summarise(N = n_distinct(date),# get number of dates
            mean = mean(DOmgL),  # get mean temperature
            SD = sd(DOmgL),      # calculate standard deviation
            SE = SD/sqrt(N))     # calculate standard error

#-Find number of instances where DO was less than 4mg/L inshore-#
ma_munson%>%       # Munson/inshore site data
  filter(DOmgL<4)  # filter to only DO measurements <4mg/L

#-Calculate diel temperature fluctuation inshore-#
ma_munson%>%                        # Munson/inshore site data
  group_by(date) %>%                # group by date
  summarise(N = n_distinct(date),   # get number of observations
    min = min(Temp_degC),           # get min Temp
            max = max(Temp_degC),   # get max Temp
            range = max - min) %>%  # get range of Temp (diel fluctuation)
  summarize(meanRange = mean(range),# get average diel fluctuation
    SD = sd(range),                 # calculate standard deviation
    SE = SD/sqrt(N))                # calculate standard error
```

## Offshore site summary statistics

```{r}
#-Calculate mean temperature +/- SE-#
ma_nursery%>%                       # Nursery/offshore site data
  filter(!is.na(maTemp_24)) %>%     # remove NAs
  summarise(N = n_distinct(date),   # get number of dates
            mean = mean(maTemp_24), # get mean temperature
            SD = sd(maTemp_24),     # calculate standard deviation
            SE = SD/sqrt(N))        # calculate standard error

#-Calculate # days where daily temperature mean was >31 degC offshore-#
ma_nursery%>%                                # Nursery/offshore site data
  ungroup()%>%                               # remove any groupings
  dplyr::group_by(date)%>%                   # group by date
  summarise(dailyMean = mean(maTemp_24))%>%  # calculate mean Temp for each date
  filter(dailyMean > 31)                     # filter to only days when mean > 31degC

#-Calculate mean DO +/- SE-#
ma_nursery%>%                    # Nursery/offshore site data
  summarise(N = n_distinct(date),# get number of dates
            mean = mean(DOmgL),  # get mean temperature
            SD = sd(DOmgL),      # calculate standard deviation
            SE = SD/sqrt(N))     # calculate standard error

#-Find number of instances where DO was less than 4mg/L offshore-#
ma_nursery%>%      # Nursery/offshore site data
  filter(DOmgL<4)  # filter to only DO measurements <4mg/L

#-Calculate diel temperature fluctuation inshore-#
ma_nursery%>%                        # Nursery/offshore site data
  group_by(date) %>%                 # group by date
  summarise(N = n_distinct(date),    # get number of observations
    min = min(Temp_degC),            # get min Temp
            max = max(Temp_degC),    # get max Temp
            range = max - min) %>%   # get range of Temp (diel fluctuation)
  summarize(meanRange = mean(range), # get average diel fluctuation
    SD = sd(range),                  # calculate standard deviation
    SE = SD/sqrt(N))                 # calculate standard error
```

# Make plots

## Set color scheme
```{r}
siteColorsPoint = c("munson" = "#f4a582", "nursery" = "#abd9e9")
siteColorsLine = c("munson" = "#d7191c", "nursery" = "#2c7bb6")
```

## Set plotting scheme
```{r}
(p <- ggplot(data = ysiSumm, aes(x = DT, color = site, fill = site)) +                              # Set up ggplot with data and aesthetic mapping
  scale_color_manual(values=siteColorsLine) +                                                       # Customize line colors using a manual scale
  scale_fill_manual(values=siteColorsPoint) +                                                       # Customize ribbon fill colors using a manual scale
  theme_bw()+                                                                                       # Set theme to black and white style
  scale_x_datetime(breaks = as.POSIXct(c("2019-07-01", "2019-08-01", "2019-09-01", "2019-10-01")),  # Customize x-axis breaks
                   labels = c( "1 Jul 2019", "1 Aug 2019", "1 Sep 2019", "1 Oct 2019")) +           # Customize x-axis labels 
  theme(axis.text.x = element_text(size = 15), axis.text.y = element_text(size = 15)) +             # Customize axis text size                                   
  theme(text=element_text(size=15)) +                                                               # Customize general text size
  theme(legend.position="none"))                                                                    # Remove legend from the plot

```


## Figure 1B - Temperature

```{r}
(bothSitesTemp <- p +                                                           # Use customized plot theme
  geom_ribbon(aes(ymin = tempMin, ymax = tempMax, alpha = 0.5), color = NA)+    # Add ribbons to represent range with transparency
  geom_line(aes (y = maTemp_24, color = site), size = 1) +                      # Add lines to represent 'maTemp_24' with specified size
    labs(x = "\nDate",                                                          # Set axis labels
       y = "Temp (\u00B0C)\n") +                                                                    
scale_y_continuous(limits = c(27.5,33), breaks = seq(27.5,33,1)))               # Set y-axis limits and breaks

#-Save plot with specified parameters directly to a specific file path-#
ggsave(file = "fig/bothSitesTemp.svg", plot = bothSitesTemp, width = 8, height = 6) # Save as scalable vector graphics file
```

##Figure 1C - DO

```{r}
#-Plot 24-hr moving average of DO over raw data-#
(bothSitesDO <- p +                                                       # Use customized plot theme
  geom_ribbon(aes(ymin = DOmin, ymax = DOmax, alpha = 0.5), color = NA) + # Add ribbons to represent range with transparency
  geom_line(aes(y = maDO_24, color = site), size = 1) +                   # Add lines to represent 'maDO_24' with specified size
  labs(x = "\nDate", y = "DO (mg/L)") +                                   # Set axis labels
  scale_y_continuous(limits = c(0,8), breaks = seq(0,8,2)))               # Set y-axis limits and breaks

#-Save plot with specified parameters directly to a specific file path-#
ggsave(file = "fig/bothSitesDO.svg", plot = bothSitesDO, width = 8, height = 6)  # Save as scalable vector graphics file
``` 

## Supplemental Figure S1

### TDS

```{r}
#-Plot 24-hr moving average of TDS over raw data-#
(bothSitesTDS <- p +                                                           # Use customized plot theme
  geom_ribbon(aes(ymin = TDSMin, ymax = TDSMax, alpha = 0.5), color = NA) +    # Add ribbons to represent range with transparency
  geom_line(aes(y = maTDS_24), size = 1) +                                     # Add lines to represent 'maTDS_24' with specified size
  labs(x = "\nDate", y = "TDS (mg/L)") +                                       # Set axis labels
  scale_y_continuous(expand = c(0,0), limits = c(34000,37000)))                # Set y-axis limits

#-Save plot with specified parameters directly to a specific file path-#
ggsave(file = "fig/bothSitesTDS.svg", plot = bothSitesTDS, width = 8, height = 6)  # Save as scalable vector graphics file
```

### pH

```{r}
#-Plot 24-hr moving average of pH over raw data-#
(bothSitespH <- p +                                                           # Use customized plot theme
  geom_ribbon(aes(ymin = pHmin, ymax = pHmax, alpha = 0.5), color = NA) +     # Add ribbons to represent range with transparency
  geom_line(aes(y = mapH_24), size = 1) +                                     # Add lines to represent 'mapH_24' with specified size
  labs(x = "\nDate", y = "pH") +                                              # Set axis labels
  scale_y_continuous(expand = c(0,0), limits = c(7.9,8.4)))                   # Set y-axis limits

#-Save plot with specified parameters directly to a specific file path-#
ggsave(file = "fig/bothSitespH.svg", plot = bothSitespH, width = 8, height = 6)  # Save as scalable vector graphics file
```

### Salinity

```{r}
#-Plot 24-hr moving average of pH over raw data-#
(bothSitesSal <- p +                                                           # Use customized plot theme
  geom_ribbon(aes(ymin = Sal_PSUmin, ymax = Sal_PSUmax, alpha = 0.5), color = NA) +     # Add ribbons to represent range with transparency
  geom_line(aes(y = maSal_24), size = 1) +                                     # Add lines to represent 'maSal_24' with specified size
  labs(x = "\nDate", y = "Salinity (psu)") +                                              # Set axis labels
  scale_y_continuous(expand = c(0,0), limits = c(34,38)))                   # Set y-axis limits

#-Save plot with specified parameters directly to a specific file path-#
ggsave(file = "fig/bothSitesSal.svg", plot = bothSitesSal, width = 8, height = 6)
```
### Comparison plot of TDS and salinity requested by handling editor

```{r}
#-Comparison plot of TDS and Salinity requested by editor-#
tds_sal_compare <- ggplot(data = ysiSumm, aes(x = maTDS_24, y = maSal_24, color = site)) +  # Set up ggplot with data and aesthetic mapping
  scale_color_manual(values=siteColorsLine) +                                               # Customize line colors using a manual scale
  geom_point(alpha = 0.01)                                                                  # Add points to the plot with transparency

#-Save plot with specified parameters directly to a specific file path-#
ggsave(file = "fig/tds_sal_compare.pdf", plot = tds_sal_compare, width = 8, height = 6) # Save as PDF
```

# Bayesian models of environmental data

## Set instructions for sampler

```{r}
n_cores <- detectCores() # Detect the number of available CPU cores
n_chains <- 2            # Set the number of chains for the sampler to run in parallel
n_iter <- 4000           # Set the total number of iterations for each chain in the sampler
n_warmup <- 2000         # Set the number of warm-up iterations, which are discarded before collecting samples
```

## Temperature models

### Visualize data

```{r}
hist(YSI_ma$maTemp_24)      # Make a histogram of the temperature data
hist(log(YSI_ma$maTemp_24)) # Make a histogram of log-transformed temperature data
```

### Define, run, and check final model

```{r}
#-Define model-#
temp_mod5 <- bf(Temp_degC ~  site + s(datetime_index, bs = "cr", k=30), 
                family =  gaussian())

#-Calculate mean of reference group for intercept parameter-#
YSI_ma_temp%>%
   group_by(site)%>%
   summarise(meanTemp = mean(Temp_degC))

#-Set priors-#
temp_prior<- c(prior(normal(0, 30.6), class = "Intercept"),
              prior(normal(0, 1), class = "b"))

#-Run model-#                          
# temp_brm5 <- brm(temp_mod5,
#                data = YSI_ma_temp,
#                prior = temp_prior,
#                cores = n_cores,
#                chains = n_chains,
#                init = "0",
#                iter = n_iter,
#                warmup = n_warmup)

#-Save model-#
# saveRDS(temp_brm5, "model outputs/temp_brm.rds")
temp_brm5<- readRDS("model outputs/temp_brm.rds")

#-Conduct posterior predictive checks-#
pp_check(temp_brm5, ndraws = 100)
pp_check(temp_brm5,type = "loo_pit_overlay", ndraws = 100)

#-Get R2 score-# 
bayes_R2(temp_brm5) 

#-Compare to next-best model using using WAIC and loo scores-#
# model_weights(temp_brm5, temp_brm4, weights = "WAIC") 
# model_weights(temp_brm5, temp_brm4, weights = "loo") 

#-Get credible intervals of final model-#
describe_posterior(temp_brm5, ci = 0.95)
describe_posterior(temp_brm5, ci = 0.65)
```

### Other temperature models compared

```{r}
# #-Define model-#
#  temp_mod1 <- bf(Temp_degC ~  site + s(datetime_index,by = site, m = 1, k=5),
#                  family =  gaussian())
# #-Run model-#
#  temp_brm1 <- brm(temp_mod1,
#                 data = YSI_ma_temp,
#                 prior = temp_prior,
#                 cores = n_cores,
#                 chains = n_chains,
#                 iter = n_iter,
#                 warmup = n_warmup)
# ###############################################################################
# #-Define model-#
#  temp_mod2 <- bf(logTemp ~  site + s(datetime_index, by = site, m=1,  k=5),
#                  family =  gaussian())
# 
# #-Run model-#
#  temp_brm2 <- brm(temp_mod2,
#                 data = YSI_ma_temp,
#                 prior = temp_prior,
#                 cores = n_cores,
#                 chains = n_chains,
#                 iter = n_iter,
#                 warmup = n_warmup)
# ###############################################################################
# #-Define model-#
#  temp_mod3 <- bf(Temp_degC ~  site + s(datetime_index, bs = "cr", k=20),
#                family =  gaussian())
# # Run model
#  temp_brm3 <- brm(temp_mod3,
#                 data = YSI_ma_temp,
#                 prior = temp_prior,
#                 cores = n_cores,
#                 chains = n_chains,
#                 init = "0",
#                 iter = n_iter,
#                 warmup = n_warmup)
# ###############################################################################
# #-Define model-#
#  temp_mod4 <- bf(Temp_degC ~  site + s(datetime_index, bs = "cr", k=50),
#                 family =  gaussian())
# #-Run model-#
#  temp_brm4 <- brm(temp_mod4,
#                 data = YSI_ma_temp,
#                 prior = temp_prior,
#                 cores = n_cores,
#                 chains = n_chains,
#                init = "0",
#                 iter = n_iter,
#                 warmup = n_warmup)
# ###############################################################################
# #-Define model-#
#  temp_mod6 <- bf(Temp_degC ~  site + s(datetime_index, by = site, id = 1, k=30) ,
#                 family =  gaussian())
# #-Run model-#
#  temp_brm6 <- brm(temp_mod6,
#                 data = YSI_ma_temp,
#                 prior = temp_prior,
#                 cores = n_cores,
#                 chains = n_chains,
#                 init = "0",
#                 iter = n_iter,
#                 warmup = n_warmup)
```

## DO models 

### Visualize data

```{r}
hist(YSI_ma$maDO_24)      # Make a histogram of the DO data
hist(log(YSI_ma$maDO_24)) # Make a histogram of the log-transformed DO data
```

### Define, run, and check final model

```{r}
#-Define model-#
do_mod6 <- bf(DOmgL ~  site + s(datetime_index, bs = "cr", k=100), 
                family =  gaussian())

#-Calculate mean of reference group for intercept parameter-#
YSI_ma_do%>%
   group_by(site)%>%
   summarise(meanDO = mean(DOmgL))

#-Set priors-#
do_prior<- c(prior(normal(6, 1), class = "Intercept"),
             prior(normal(0, 2), class = "b"))

#-Run model-#
# do_brm6 <- brm(do_mod6,
#                data = YSI_ma_do,
#                prior = do_prior,
#                cores = n_cores,
#                chains = n_chains,
#                init = "0",
#                iter = n_iter,
#                warmup = n_warmup)

#-Save model-#
# saveRDS(do_brm6, "model outputs/DO_brm.rds")
do_brm6 <- readRDS("model outputs/DO_brm.rds")

#-Conduct posterior predictive checks-#
pp_check(do_brm6, ndraws = 100)
pp_check(do_brm6,type = "loo_pit_overlay", ndraws = 100)

#-Get R2 score-# 
bayes_R2(do_brm6) 

#-Compare to next-best model using using WAIC and loo scores-#
# model_weights(do_brm5, do_brm6, weights = "WAIC")
# model_weights(do_brm5, do_brm6, weights = "loo")

#-Get credible intervals of best model-#
describe_posterior(do_brm6, ci = 0.95)
describe_posterior(do_brm6, ci = 0.65)
```


### Other DO models compared

```{r}
# # Define model
# do_mod1 <- bf(DOmgL ~  site + s(datetime_index,by = site, m = 1, k=5), 
#                 family =  gaussian())
# 
# # Run model                          
#  do_brm1 <- brm(do_mod1,
#                 data = YSI_ma_do,
#                 prior = do_prior,
#                 cores = n_cores,
#                 chains = n_chains,
#                 iter = n_iter,
#                 warmup = n_warmup)
################################################################################
# # Define model
# do_mod2 <- bf(DOmgL ~  site + s(datetime_index, bs = "cr", k=20), 
#                 family =  gaussian())
# 
# # Run model                           
#  do_brm2 <- brm(do_mod2,
#                 data = YSI_ma_do,
#                 prior = do_prior,
#                 cores = n_cores,
#                 chains = n_chains,
#                init = "0",
#                 iter = n_iter,
#                 warmup = n_warmup)
################################################################################
# # Define model
#  do_mod3 <- bf(DOmgL ~  site + s(datetime_index, bs = "cr", k=30), 
#                 family =  gaussian())
# # Run model                              
#  do_brm3 <- brm(do_mod3,
#                 data = YSI_ma_do,
#                 prior = do_prior,
#                 cores = n_cores,
#                 chains = n_chains,
#                init = "0",
#                 iter = n_iter,
#                 warmup = n_warmup)
################################################################################
# # Define model
# do_mod4 <- bf(DOmgL ~  site + s(datetime_index, bs = "cr", k=40), 
#                 family =  gaussian())
# 
# # Run model                             
#  do_brm4 <- brm(do_mod4,
#                 data = YSI_ma_do,
#                 prior = do_prior,
#                 cores = n_cores,
#                 chains = n_chains,
#                 init = "0",
#                 iter = n_iter,
#                 warmup = n_warmup)
################################################################################
# # Define model
# do_mod5 <- bf(DOmgL ~  site + s(datetime_index, bs = "cr", k=50), 
#                 family =  gaussian())
# 
# # Run model   
#  do_brm5 <- brm(do_mod5,
#                 data = YSI_ma_do,
#                 prior = do_prior,
#                 cores = n_cores,
#                 chains = n_chains,
#                init = "0",
#                 iter = n_iter,
#                 warmup = n_warmup)
################################################################################
# # Define model
# do_mod7 <- bf(DOmgL ~  site , 
#                 family =  gaussian())
# # Run model
#  do_brm7 <- brm(do_mod7,
#                 data = YSI_ma_do,
#                 prior = do_prior,
#                 cores = n_cores,
#                 chains = n_chains,
#                 iter = n_iter,
#                 warmup = n_warmup)
```
